Change entity IDs from int to long and rename IAggregateEntity to IEntity
Changed all entity IDs from int (4 bytes, 2.1 billion limit) to long (8 bytes, 9.2 quintillion limit) to eliminate ID exhaustion concerns with minimal performance cost. This prevents issues with deleted records consuming ID space and provides infinite runway for high-churn scenarios.

Renamed IAggregateEntity to IEntity since this interface applies to all persisted entities, not just DDD aggregate roots. This better reflects the pragmatic layered architecture without implying full DDD implementation.

Added IEntity constraint to Repository<T> for type safety - only entities implementing IEntity can use the repository pattern.

Migration updates User table ID from integer to bigint and adds UpdatedAt column.

+++ PagePlay.Site/Application/Accounts.Domain/Repository/IUserRepository.cs
@@ -6,6 +6,6 @@ namespace PagePlay.Site.Application.Accounts.Domain.Repository;
 public interface IUserRepository : IRepository<User>
 {
     Task<User> GetByEmailAsync(string email);
-    Task<User> GetByIdAsync(int id);
+    Task<User> GetByIdAsync(long id);
     Task<bool> EmailExistsAsync(string email);
 }
